segfult

+++ 0x05-pointers_arrays_strings/4-print_rev.c
@@ -9,12 +9,13 @@
  */
 void print_rev(char *s)
 {
-	short int len;
+	int len;
 
 	len = 0;
 	while (s[len] != '\0')
 		len++;
-	len--; /*removing the extra char that rep '\0'*/
+
+	len--;
 	while (len >= 0)
 	{
 		_putchar(s[len]);
